Board와 Comment 각각 컨트롤러, 서비스 분리 및 멤버 널체크 삭제
Board, Comment 각각의 컨트롤러와 서비스 분리, 그에 따른 패키지 구조 조정
DTO는 컨트롤러에서 완성해 서비스로 넘기기로 통일
서비스에서는 항상 엔티티를 영속화 하고 작업을 수행

댓글에 대한 CRUD는 게시글을 통해 수행되도록 변경
읽기 : board.getComment()
생성 : board.addComment()
수정 : comment.update()
삭제 : board.deleteComment()

--

댓글 수정 Dto에 익명여부 추가
