--- src/main/java/hello/kssoftware/board/Board.java
@@ -1,45 +1,0 @@
-package hello.kssoftware.board;
-
-import hello.kssoftware.login.Member;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-import org.hibernate.annotations.ColumnDefault;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Entity
-@Getter
-@Setter
-public class Board {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "user_id")
-    private Member writer;
-
-    private String title;
-
-    private String content;
-
-    private LocalDateTime createDate;
-
-    private LocalDateTime updateDate;
-
-    @ColumnDefault(value = "0")
-    private Integer views = 0;
-
-    @OneToMany(mappedBy = "board", fetch = FetchType.EAGER, cascade = CascadeType.REMOVE)
-    private List<Comment> comments;
-
-    public Comment getComment(Long commentId) {
-        return comments.stream()
-                .filter(comment -> comment.getId().equals(commentId))
-                .findFirst()
-                .orElse(null);
-    }
-}
