+++ src/main/java/hello/kssoftware/board/dto/CommentUpdateDto.java
@@ -8,7 +8,9 @@ import lombok.Setter;
 @Setter
 public class CommentUpdateDto {
 
+    private Long boardId;
     private Long commentId;
     @NotEmpty(message = "내용을 입력하세요.")
     private String content;
+    private Boolean anonymousYn;
 }
